requirements.txt
fastapi>=0.115.8,<0.116.0
fastapi-mail>=1.4.2, <1.5.0
fastapi-pagination>=0.12.31, <0.13.0
pydantic>=2.10.6,<2.11.0
sqlalchemy>=2.0.37,<2.1.0
psycopg2>=2.9.10,<2.10.0
sqladmin>=0.20.1, <0.21.0
alembic>=1.14.1, <1.15.0
passlib[bcrypt]>=1.7.4,<2.0.0
pytest>=8.3.4,<8.4.0
Faker>=30.8.2,<31.0.0
PyJWT>=2.10.1,<2.11.0
uvicorn==0.30.3
